separate body and resutantcard
