+++ src/App.js
@@ -1,7 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import Header from '../src/component/Header'
-
+import Header from './component/Header'
+import Body from './component/Body';
+import ResturantCard from "./component/ResturantCard";
 const restaurantList = [
   {
     type: "restaurant",
@@ -1820,44 +1821,8 @@ const restaurantList = [
   },
 ];
 
-//ResturantCard
-//we can also desctructure the props name as({resname,cuisine})
-const ResturantCard = (props) => {
-  //console.log(props);
-  const { resData } = props;
-  //help of optional chaining
-  const{cloudinaryImageId,name,cuisines,avgRating,costForTwo,deliveryTime}=resData?.data;
-  return (
-    <div className="res-card">
-      <img
-        className="res-logo"
-        alt="res-logo"
-        src={"https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_508,h_320,c_fill/"+cloudinaryImageId}
-      />
-      <h1>{name}</h1>
-      <h4>{cuisines.join(", ")}</h4>
-      <h4>{avgRating} Stars</h4>
-      <h4>{costForTwo / 100} </h4>
-      <h4>{deliveryTime}minutes</h4>
-    </div>
-  );
-};
 
-//Body
-const Body = () => {
-  return (
-    <div className="body">
-      <div className="search">Search</div>
-      <div className="res-container">
-       {
-         restaurantList.map((restuarant)=>(
-          <ResturantCard key={restuarant.data.id} resData={restuarant}/>
-         ))
-       }
-      </div>
-    </div>  
-  );
-};
+
 
 const AppLayout = () => {
   return (
